# orchestrator/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Распределённый калькулятор",
        "title": "yandex_calculator",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/c_resources": {
            "get": {
                "description": "Выводит список вычислительных мощностей с задачами, которые на них выполняются.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "computing_resources"
                ],
                "summary": "Выводит список вычислительных мощностей с задачами, которые на них выполняются.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.ComputingResource"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/calculate": {
            "post": {
                "description": "Добавляет новую задачу и отдаёт её на обработку.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Добавляет новую задачу и отдаёт её на обработку.",
                "parameters": [
                    {
                        "description": "Задача",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.AddTaskReqBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/operations": {
            "get": {
                "description": "Выводит список доступных операций и время их работы.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Выводит список доступных операций.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Operation"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет время указанной операции.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "operations"
                ],
                "summary": "Обновляет время указанной операции.",
                "parameters": [
                    {
                        "description": "оператор с временем его выполнения",
                        "name": "operation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/operation.UpdateOperationTimeReqBody"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/results/{task_id}": {
            "get": {
                "description": "Получает результат по идентификатору задачи.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получает результат по идентификатору задачи.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор задачи",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tasks/{task_id}": {
            "get": {
                "description": "Получает задачу по её идентификатору.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Получает задачу по её идентификатору.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор задачи",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.ComputingResource": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "occupied": {
                    "type": "boolean"
                },
                "task": {
                    "type": "string"
                }
            }
        },
        "entity.ExpressionStatus": {
            "type": "string",
            "enum": [
                "error",
                "in_progress",
                "ok"
            ],
            "x-enum-varnames": [
                "Error",
                "InProgress",
                "OK"
            ]
        },
        "entity.Operation": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entity.Task": {
            "type": "object",
            "properties": {
                "calculatedAt": {
                    "type": "string"
                },
                "calculatedBy": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "expression": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/entity.ExpressionStatus"
                }
            }
        },
        "operation.UpdateOperationTimeReqBody": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "task.AddTaskReqBody": {
            "type": "object",
            "properties": {
                "expression": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        }
    }
}